Reorganize boilerplate
