+++ src/rust/async/future.rs
@@ -1,15 +1,7 @@
 use std::future::Future;
 use std::pin::Pin;
 use std::task::Context;
-use std::task::Poll::Pending;
-use std::task::Poll::Ready;
-use std::task::Waker;
-
-use cxx::ExternType;
-
-use crate::ffi::KjWaker;
-
-use crate::Result;
+use std::task::Poll;
 
 // Expose Pin<Box<dyn Future<Output = ()>> to C++ as BoxFutureVoid.
 //
@@ -21,7 +13,12 @@ use crate::Result;
 
 pub struct BoxFuture<T>(Pin<Box<dyn Future<Output = T> + Send>>);
 
-// TODO(now): Might as well implement Future for BoxFuture<T>.
+impl<T> Future for BoxFuture<T> {
+    type Output = T;
+    fn poll(mut self: Pin<&mut Self>, cx: &mut Context) -> Poll<T> {
+        self.0.as_mut().poll(cx)
+    }
+}
 
 // A From implementation to make it easier to convert from an arbitrary Future
 // type into a BoxFuture<T>.
@@ -39,125 +36,8 @@ impl<T, F: Future<Output = T> + Send + 'static> From<Pin<Box<F>>> for BoxFuture<
 #[repr(transparent)]
 pub struct PtrBoxFuture<T>(*mut BoxFuture<T>);
 
-// =======================================================================================
-
-// TODO(now): Define these trait implementations with a macro?
-
-// Safety: The size of a Pin<P> is the size of P; the size of a Box<T> is the size of a reference to
-// T, and references to `dyn Trait` types contain two pointers: one for the object, one for the
-// vtable. So, the size of `Pin<Box<dyn Future<Output = T>>>` (a.k.a. `BoxFuture<T>`) is two
-// pointers, and that is unlikely to change.
-//
-// https://doc.rust-lang.org/std/keyword.dyn.html
-// - "As such, a dyn Trait reference contains two pointers."
-unsafe impl ExternType for BoxFuture<()> {
-    type Id = cxx::type_id!("workerd::rust::async::BoxFutureVoid");
-    type Kind = cxx::kind::Trivial;
-}
-
-// Safety: Raw pointers are the same size in both languages.
-unsafe impl ExternType for PtrBoxFuture<()> {
-    type Id = cxx::type_id!("workerd::rust::async::PtrBoxFutureVoid");
-    type Kind = cxx::kind::Trivial;
-}
-
-use crate::ffi::BoxFutureFulfillerVoid;
-
-pub fn box_future_poll_void(
-    future: &mut BoxFuture<()>,
-    waker: &KjWaker,
-    fulfiller: Pin<&mut BoxFutureFulfillerVoid>,
-) -> bool {
-    let waker = Waker::from(waker);
-    let mut cx = Context::from_waker(&waker);
-    match future.0.as_mut().poll(&mut cx) {
-        Ready(_v) => {
-            fulfiller.fulfill();
-            true
-        }
-        Pending => false,
+impl<T> PtrBoxFuture<T> {
+    pub unsafe fn drop_in_place(self) {
+        std::ptr::drop_in_place(self.0);
     }
-}
-
-pub unsafe fn box_future_drop_in_place_void(ptr: PtrBoxFuture<()>) {
-    std::ptr::drop_in_place(ptr.0);
-}
-
-// ---------------------------------------------------------
-
-// Safety: The size of a Pin<P> is the size of P; the size of a Box<T> is the size of a reference to
-// T, and references to `dyn Trait` types contain two pointers: one for the object, one for the
-// vtable. So, the size of `Pin<Box<dyn Future<Output = T>>>` (a.k.a. `BoxFuture<T>`) is two
-// pointers, and that is unlikely to change.
-//
-// https://doc.rust-lang.org/std/keyword.dyn.html
-// - "As such, a dyn Trait reference contains two pointers."
-unsafe impl ExternType for BoxFuture<Result<()>> {
-    type Id = cxx::type_id!("workerd::rust::async::BoxFutureFallibleVoid");
-    type Kind = cxx::kind::Trivial;
-}
-
-// Safety: Raw pointers are the same size in both languages.
-unsafe impl ExternType for PtrBoxFuture<Result<()>> {
-    type Id = cxx::type_id!("workerd::rust::async::PtrBoxFutureFallibleVoid");
-    type Kind = cxx::kind::Trivial;
-}
-
-use crate::ffi::BoxFutureFulfillerFallibleVoid;
-
-pub fn box_future_poll_fallible_void(
-    future: &mut BoxFuture<Result<()>>,
-    waker: &KjWaker,
-    fulfiller: Pin<&mut BoxFutureFulfillerFallibleVoid>,
-) -> Result<bool> {
-    let waker = Waker::from(waker);
-    let mut cx = Context::from_waker(&waker);
-    match future.0.as_mut().poll(&mut cx) {
-        Ready(Ok(_v)) => {
-            fulfiller.fulfill();
-            Ok(true)
-        }
-        Ready(Err(e)) => Err(e),
-        Pending => Ok(false),
-    }
-}
-
-pub unsafe fn box_future_drop_in_place_fallible_void(ptr: PtrBoxFuture<Result<()>>) {
-    std::ptr::drop_in_place(ptr.0);
-}
-
-// ---------------------------------------------------------
-
-unsafe impl ExternType for BoxFuture<Result<i32>> {
-    type Id = cxx::type_id!("workerd::rust::async::BoxFutureFallibleI32");
-    type Kind = cxx::kind::Trivial;
-}
-
-// Safety: Raw pointers are the same size in both languages.
-unsafe impl ExternType for PtrBoxFuture<Result<i32>> {
-    type Id = cxx::type_id!("workerd::rust::async::PtrBoxFutureFallibleI32");
-    type Kind = cxx::kind::Trivial;
-}
-
-use crate::ffi::BoxFutureFulfillerFallibleI32;
-
-pub fn box_future_poll_fallible_i32(
-    future: &mut BoxFuture<Result<i32>>,
-    waker: &KjWaker,
-    fulfiller: Pin<&mut BoxFutureFulfillerFallibleI32>,
-) -> Result<bool> {
-    let waker = Waker::from(waker);
-    let mut cx = Context::from_waker(&waker);
-    match future.0.as_mut().poll(&mut cx) {
-        Ready(Ok(v)) => {
-            fulfiller.fulfill(v);
-            Ok(true)
-        }
-        Ready(Err(e)) => Err(e),
-        Pending => Ok(false),
-    }
-}
-
-pub unsafe fn box_future_drop_in_place_fallible_i32(ptr: PtrBoxFuture<Result<i32>>) {
-    std::ptr::drop_in_place(ptr.0);
 }
